Module rs_utils::array

Array utilities

Functions

fst
Create a new array from the first elements of an array of pairs of the same length as the first array.

fst_default
Create a new array from the first elements of an array of pairs or the default if the destination array length is longer than the input array length.

fst_initial
Create a new array from the first elements of an array of pairs of equal or greater length than the first.

snd
Create a new array from the second elements of an array of pairs of the same length as the second array.

snd_default
Create a new array from the second elements of an array of pairs or the default if the destination array length is longer than the input array length.

snd_initial
Create a new array from the second elements of an array of pairs of equal or greater length than the second.