Function rs_utils::array::fst_default

pub fn fst_default<A, B, N, M>(
    from: &GenericArray<(A, B), N>
) -> GenericArray<A, M> where
    A: Clone + Default,
    N: ArrayLength<(A, B)>,
    M: ArrayLength<A>, 

Create a new array from the first elements of an array of pairs or the default if the destination array length is longer than the input array length.

use generic_array::GenericArray;
use typenum::consts::*;
let xs : [(u8, f32); 2]        = [(5, 1.0), (100, -1.0)];
let ys : GenericArray <u8, U3> = fst_default (&xs.into());
assert_eq!(ys.as_slice(), [5, 100, 0]);